sec try update commit

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: mchergui <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/11 18:15:39 by mchergui          #+#    #+#             */
-/*   Updated: 2019/10/23 12:45:39 by mchergui         ###   ########.fr       */
+/*   Updated: 2019/10/26 19:35:20 by mchergui         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (str[i] != '\0')
